seperated: settings dropdown comp. from header
