+++ src/layouts/AppLayout.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { AppLayoutProps } from "@src/Interfaces/ILayouts";
 import BottomNavbar from "@components/BottomNavbar/BottomNavbar";
 
-import Header from "../common/Header/Header";
+import Header from "../components/Header/Header";
 
 const AppLayout: React.FC<AppLayoutProps> = ({ children, title, debounceSearch }) => (
   <>
